+++ src/components/CommitDetails/index.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { BranchType, CommitNode } from "../../helpers/types";
+import { Branch, CommitNode } from "../../helpers/types";
 import BranchLabel from "../BranchLabel";
 import { excerpt } from "../Curves/utils";
 import css from "./index.module.css";
@@ -7,7 +7,7 @@ import { Tooltip } from "react-tooltip";
 
 type Props = {
   commit: CommitNode;
-  branch: BranchType[];
+  branch: Branch[];
   mouseOver: () => void;
   mouseLeave: () => void;
 };
@@ -18,7 +18,7 @@ export default function CommitDetails({
   mouseLeave,
   mouseOver,
 }: Props) {
-  const date = new Date(commit.committerDate).toLocaleDateString();
+  const date = commit.commitDate.toLocaleDateString();
   const hashBr = commit.hash.slice(0, 7);
   const committer = commit.committer;
   const message = commit.message || "";
@@ -45,10 +45,7 @@ export default function CommitDetails({
         ) : (
           <span className={css.bold}>{commitHashAuthorDate}</span>
         )}
-        <BranchLabel
-                 branchColor={commit.commitColor}
-                branches={branch}
-               /> 
+        <BranchLabel branchColor={commit.commitColor} branches={branch} />
       </div>
       <div
         data-tooltip-content={message}
@@ -59,7 +56,11 @@ export default function CommitDetails({
         {excerpt(message, 80)}
       </div>
       {message.length > 80 && (
-        <Tooltip id={`commit-${commit.hash}-msg`} className={css.tooltip} noArrow/>
+        <Tooltip
+          id={`commit-${commit.hash}-msg`}
+          className={css.tooltip}
+          noArrow
+        />
       )}
     </div>
   );