+++ src/components/CommitGraph/index.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
-import { BranchType, Commit, GraphStyle } from "../../helpers/types";
+import { Branch, Commit, GraphStyle } from "../../helpers/types";
 import {
   defaultStyle,
-  getCommits,
+  formatCommits,
   setBranchAndCommitColor,
 } from "../../helpers/utils";
 import Branches from "../Branches";
@@ -12,10 +12,11 @@ import { getCommitDotPosition } from "../CommitDot/utils";
 import Curves from "../Curves";
 import { computePosition } from "./computePosition";
 import css from "./index.module.css";
+import WithPagination from "./WithPagination";
 
 export type Props = {
   commits: Commit[];
-  branchHeads: BranchType[];
+  branchHeads: Branch[];
   graphStyle?: GraphStyle;
 };
 
@@ -27,7 +28,7 @@ export default function CommitGraph({
   const [showBlock, setShowBlock] = useState(false);
   const [topPos, setTopPos] = useState(0);
 
-  const commitNodes = getCommits(commits);
+  const commitNodes = formatCommits(commits);
   const { commitSpacing, branchSpacing, branchColors, nodeRadius } = {
     ...defaultStyle,
     ...graphStyle,
@@ -90,7 +91,7 @@ export default function CommitGraph({
             commit,
           );
           const branch = branchHeads.filter(
-            b => b.headCommitHash === commit.hash,
+            b => b.commit.sha === commit.hash,
           );
           const mouseOver = () => {
             setShowBlock(true);
@@ -139,3 +140,5 @@ function getCommitInfoLeftPosition(width: number) {
   }
   return 510;
 }
+
+CommitGraph.WithPagination = WithPagination;